--- domain/generator/src/main/kotlin/com/few/generator/service/DateRange.kt
@@ -1,8 +1,0 @@
-package com.few.generator.service
-
-import java.time.LocalDateTime
-
-data class DateRange(
-    val start: LocalDateTime,
-    val end: LocalDateTime,
-)
