+++ domain/generator/src/main/kotlin/com/few/generator/service/implement/NewsletterTemplateData.kt
@@ -1,4 +1,4 @@
-package com.few.generator.service
+package com.few.generator.service.implement
 
 import com.few.common.domain.Category
 import com.few.email.GenData
